pyproject.toml
[tool.poetry]
name = "cashback-api"
version = "0.1.0"
description = ""
authors = ["Marcos Paulo <marcosbarretopaulo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
psycopg2-binary = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.4"}
pydantic-settings = "^2.7.0"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
